fixed precision of zero point

+++ torch2trt/contrib/qat/layers/quant_conv.py
@@ -128,7 +128,7 @@ class QuantConv2d(_QuantConvNd):
         else:
             quant_input = torch.fake_quantize_per_channel_affine(input,
                     quantizer.quant_scale,
-                    quantizer.zero_point,
+                    quantizer.zero_point.to(torch.long),
                     quantizer.quant_axis.to(torch.int32).item(),
                     quantizer.quant_min.to(torch.long).item(),
                     quantizer.quant_max.to(torch.long).item())
